+++ cmd/environments/get_test.go
@@ -1,10 +1,8 @@
 package environments_test
 
 import (
-	"encoding/json"
-	"fmt"
 	"ldcli/cmd"
-	"strings"
+	"ldcli/cmd/output"
 	"testing"
 
 	ldapi "github.com/launchdarkly/api-client-go/v14"
@@ -172,9 +170,7 @@ func TestOutputFlagGet(t *testing.T) {
 			"tags": null
 		}`
 
-		output, err := CmdOutput(outputFlag, EnvironmentOutputter{
-			environment: environment,
-		})
+		output, err := output.CmdOutput(outputFlag, environments.NewEnvironmentOutputter(environment))
 
 		require.NoError(t, err)
 		assert.JSONEq(t, expected, output)
@@ -184,9 +180,7 @@ func TestOutputFlagGet(t *testing.T) {
 		outputFlag := "plaintext"
 		expected := "test-name (test-key)"
 
-		output, err := CmdOutput(outputFlag, EnvironmentOutputter{
-			environment: environment,
-		})
+		output, err := output.CmdOutput(outputFlag, environments.NewEnvironmentOutputter(environment))
 
 		require.NoError(t, err)
 		assert.Equal(t, expected, output)
@@ -196,59 +190,11 @@ func TestOutputFlagGet(t *testing.T) {
 		outputFlag := ""
 		expected := "test-name (test-key)"
 
-		output, err := CmdOutput(outputFlag, EnvironmentOutputter{
-			environment: environment,
-		})
+		output, err := output.CmdOutput(outputFlag, environments.NewEnvironmentOutputter(environment))
 
 		require.NoError(t, err)
 		assert.Equal(t, expected, output)
 	})
 
 	t.Run("when flag is invalid", func(t *testing.T) {})
-}
-
-type Outputter interface {
-	JSON() (string, error)
-	String() string
-}
-
-type EnvironmentOutputter struct {
-	environment *ldapi.Environment
-}
-
-func (o EnvironmentOutputter) JSON() (string, error) {
-	responseJSON, err := json.Marshal(o.environment)
-	if err != nil {
-		return "", err
-	}
-
-	return string(responseJSON), nil
-}
-
-func (o EnvironmentOutputter) String() string {
-	fnPlaintext := func(p *ldapi.Environment) string {
-		return fmt.Sprintf("%s (%s)", p.Name, p.Key)
-	}
-
-	return formatColl([]*ldapi.Environment{o.environment}, fnPlaintext)
-}
-
-func CmdOutput(outputKind string, outputter Outputter) (string, error) {
-	switch outputKind {
-	case "json":
-		return outputter.JSON()
-	case "plaintext":
-		return outputter.String(), nil
-	default:
-		return outputter.String(), nil
-	}
-}
-
-func formatColl[T any](coll []T, fn func(T) string) string {
-	lst := make([]string, 0, len(coll))
-	for _, c := range coll {
-		lst = append(lst, fn(c))
-	}
-
-	return strings.Join(lst, "\n")
 }
