+++ internal/environments/client.go
@@ -2,11 +2,12 @@ package environments
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
-	"strings"
 
 	ldapi "github.com/launchdarkly/api-client-go/v14"
 
+	"ldcli/cmd/output"
 	"ldcli/internal/client"
 	"ldcli/internal/errors"
 )
@@ -41,35 +42,38 @@ func (c EnvironmentsClient) Get(
 
 	}
 
-	// outputType := NewOutputType(viper.GetString("output"))
-	// output, err := CmdOutput(outputType, EnvironmentOutputter{
-	// 	environment: environment,
-	// })
-	// if err != nil {
-	// 	return nil, errors.NewLDAPIError(err)
+	output, err := output.CmdOutput("", NewEnvironmentOutputter(environment))
+	if err != nil {
+		return nil, errors.NewLDAPIError(err)
 
-	// }
+	}
 
-	// return output
+	return []byte(output), nil
+}
 
+type EnvironmentOutputter struct {
+	environment *ldapi.Environment
+}
+
+func (o EnvironmentOutputter) JSON() (string, error) {
+	responseJSON, err := json.Marshal(o.environment)
+	if err != nil {
+		return "", err
+	}
+
+	return string(responseJSON), nil
+}
+
+func (o EnvironmentOutputter) String() string {
 	fnPlaintext := func(p *ldapi.Environment) string {
 		return fmt.Sprintf("%s (%s)", p.Name, p.Key)
 	}
-	return foo([]*ldapi.Environment{environment}, fnPlaintext), nil
 
-	// responseJSON, err := json.Marshal(environment)
-	// if err != nil {
-	// 	return nil, err
-	// }
-
-	// return responseJSON, nil
+	return output.FormatColl([]*ldapi.Environment{o.environment}, fnPlaintext)
 }
 
-func foo[T any](coll []T, fn func(T) string) []byte {
-	lst := make([]string, 0, len(coll))
-	for _, c := range coll {
-		lst = append(lst, fn(c))
+func NewEnvironmentOutputter(environment *ldapi.Environment) EnvironmentOutputter {
+	return EnvironmentOutputter{
+		environment: environment,
 	}
-
-	return []byte(strings.Join(lst, "\n"))
 }
